+++ PKHeX.Core/Legality/Encounters/Templates/Gen7/EncounterTransfer7.cs
@@ -14,10 +14,10 @@ public sealed record EncounterTransfer7(GameVersion Version, ushort Species, byt
     public EntityContext Context => EntityContext.Gen7;
     public Ball FixedBall => Ball.Poke;
     public bool EggEncounter => false;
-    public int EggLocation => 0;
+    public ushort EggLocation => 0;
 
     public byte Form => 0;
-    public int Location { get; private init; }
+    public ushort Location { get; private init; }
     public Shiny Shiny { get; private init; }
     public AbilityPermission Ability { get; private init; }
     public bool FatefulEncounter { get; private init; }
